refactor: loosening dependencies
