+++ src/commands/setting.ts
@@ -1,6 +1,7 @@
 /* eslint-disable new-cap */
 import {ProjectSettings} from '../dotfile';
-import {ERROR, getProjectSettings, logError, saveProject} from '../utils';
+import {ERROR} from '../messages';
+import {getProjectSettings, logError, saveProject} from '../utils';
 
 /**
  * Gets or sets a setting in .clasp.json
