+++ src/commands/status.ts
@@ -1,6 +1,7 @@
 import {getProjectFiles, logFileList} from '../files';
 import {isValidManifest} from '../manifest';
-import {checkIfOnline, getProjectSettings, LOG} from '../utils';
+import {LOG} from '../messages';
+import {checkIfOnline, getProjectSettings} from '../utils';
 
 /**
  * Displays the status of which Apps Script files are ignored from .claspignore
