+++ src/constants.ts
@@ -1,0 +1,7 @@
+// @see https://nodejs.org/api/fs.html#fs_fs_readfilesync_path_options
+export const FS_OPTIONS = {encoding: 'utf8'};
+
+// Names / Paths
+export const PROJECT_NAME = 'clasp';
+export const PROJECT_MANIFEST_BASENAME = 'appsscript';
+export const PROJECT_MANIFEST_FILENAME = `${PROJECT_MANIFEST_BASENAME}.json`;
