+++ test/commands/clone.ts
@@ -5,7 +5,7 @@ import fs from 'fs-extra';
 import {after, before, describe, it} from 'mocha';
 
 import {URL} from '../../src/urls';
-import {ERROR, LOG} from '../../src/utils';
+import {ERROR, LOG} from '../../src/messages';
 import {CLASP, SCRIPT_ID} from '../constants';
 import {cleanup, setup} from '../functions';
