+++ test/commands/default.ts
