+++ test/commands/open.ts
@@ -4,7 +4,7 @@ import {spawnSync} from 'child_process';
 import {after, before, describe, it} from 'mocha';
 
 import {URL} from '../../src/urls';
-import {ERROR, LOG} from '../../src/utils';
+import {ERROR, LOG} from '../../src/messages';
 import {CLASP, PARENT_ID, PROJECT_ID, SCRIPT_ID} from '../constants';
 import {cleanup, setup} from '../functions';
